Fixes #15
Fixes a bug where autocomplete does not work properly in editor panes
(as opposed to the REPL). This commit also makes a few other small
improvements, such as updated documentation so that tab complete doesn't
interfere with Lisp indentation, and a print out that it's trying to
connect (can be useful if connecting takes a few seconds).